fix of word error michmash

+++ ai_Model/mask_utils.py
@@ -3,4 +3,4 @@ import torch
 
 def generate_square_subsequent_mask(sz):
     mask = torch.triu(torch.ones(sz, sz), diagonal=1).bool()
-    return mask
+    return mask.float()
